Backend/requirements.txt
Django==4.2.5
djangorestframework
djoser
mysqlclient